# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: Main service API
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
...
  /admin/users:
    get:
      description: Возвращает информацию обо всех пользователях (учитываются параметры
        ограничения выборки), либо о конкретных (учитываются указанные идентификаторы)
      operationId: getUsers
      parameters:
      - description: id пользователей
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - description: количество элементов, которые нужно пропустить для формирования
          текущего набора
        explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          type: integer
        style: form
      - description: количество элементов в наборе
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserDto'
                type: array
          description: Пользователи найдены
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Получение информации о пользователях
      tags:
      - 'Admin: Пользователи'
      x-accepts: application/json
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
        description: данные добавляемого пользователя
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
          description: Пользователь зарегистрирован
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Добавление нового пользователя
      tags:
      - 'Admin: Пользователи'
      x-contentType: application/json
      x-accepts: application/json
  /admin/users/{userId}:
    delete:
      operationId: delete
      parameters:
      - description: id пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Пользователь удален
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Удаление пользователя
      tags:
      - 'Admin: Пользователи'
      x-accepts: application/json
...
  /users/{userId}/requests:
    get:
      operationId: getUserRequests
      parameters:
      - description: id текущего пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ParticipationRequestDto'
                type: array
          description: Найдены запросы на участие
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Получение информации о заявках текущего пользователя на участие в чужих
        событиях
      tags:
      - 'Private: Запросы на участие'
      x-accepts: application/json
    post:
      description: |-
        Обратите внимание:
        - нельзя добавить повторный запрос
        - инициатор события не может добавить запрос на участие в своём событии
        - нельзя участвовать в неопубликованном событии
        - если у события достигнут лимит запросов на участие - необходимо вернуть ошибку
        - если для события отключена пре-модерация запросов на участие, то запрос должен автоматически перейти в состояние подтвержденного
      operationId: addParticipationRequest
      parameters:
      - description: id текущего пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id события
        explode: true
        in: query
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationRequestDto'
          description: Заявка создана
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Добавление запроса от текущего пользователя на участие в событии
      tags:
      - 'Private: Запросы на участие'
      x-accepts: application/json
  /users/{userId}/requests/{requestId}/cancel:
    patch:
      operationId: cancelRequest
      parameters:
      - description: id текущего пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id запроса на участие
        explode: false
        in: path
        name: requestId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationRequestDto'
          description: Заявка отменена
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Отмена своего запроса на участие в событии
      tags:
      - 'Private: Запросы на участие'
      x-accepts: application/json
components:
  schemas:
...
    ApiError:
      description: Сведения об ошибке
      properties:
        errors:
          description: Список стектрейсов или описания ошибок
          example: []
          items:
            description: Список стектрейсов или описания ошибок
            example: '[]'
            type: string
          type: array
        message:
          description: Сообщение об ошибке
          example: Only pending or canceled events can be changed
          type: string
        reason:
          description: Общее описание причины ошибки
          example: For the requested operation the conditions are not met.
          type: string
        status:
          description: Код статуса HTTP-ответа
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: FORBIDDEN
          type: string
        timestamp:
          description: Дата и время когда произошла ошибка (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: 2022-06-09 06:27:23
          type: string
      type: object
...
    NewUserRequest:
      description: Данные нового пользователя
      example:
        name: Иван Петров
        email: ivan.petrov@practicummail.ru
      properties:
        email:
          description: Почтовый адрес
          example: ivan.petrov@practicummail.ru
          type: string
        name:
          description: Имя
          example: Иван Петров
          type: string
      required:
      - email
      - name
      type: object
    ParticipationRequestDto:
      description: Заявка на участие в событии
      example:
        requester: 2
        created: 2022-09-06T21:10:05.432
        id: 3
        event: 1
        status: PENDING
      properties:
        created:
          description: Дата и время создания заявки
          example: 2022-09-06T21:10:05.432
          type: string
        event:
          description: Идентификатор события
          example: 1
          format: int64
          type: integer
        id:
          description: Идентификатор заявки
          example: 3
          format: int64
          type: integer
        requester:
          description: Идентификатор пользователя, отправившего заявку
          example: 2
          format: int64
          type: integer
        status:
          description: Статус заявки
          example: PENDING
          type: string
      type: object
...
    UserDto:
      description: Пользователь
      example:
        name: Петров Иван
        id: 1
        email: petrov.i@practicummail.ru
      properties:
        email:
          description: Почтовый адрес
          example: petrov.i@practicummail.ru
          type: string
        id:
          description: Идентификатор
          example: 1
          format: int64
          readOnly: true
          type: integer
        name:
          description: Имя
          example: Петров Иван
          type: string
      required:
      - email
      - name
      type: object
...